% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRiskScore.R
\name{predRiskScore}
\alias{predRiskScore}
\title{Score trajectory cluster model}
\usage{
predRiskScore(input = NULL)
}
\arguments{
\item{age}{patient age}

\item{sex}{patient gender}

\item{time}{follow up time from surgery}

\item{acei}{ace inhibitor(drug). ACE inhibitors treat a variety of conditions, 
such as high blood pressure, scleroderma and migraines.}

\item{dm}{preoperative diabetes}

\item{creat}{preoperative serum creatinine}

\item{hc}{preoperative high cholesterol}

\item{prenyha}{preoperative New York Heart Association (NYHA): 1 = I/II; 3 = III/IV}

\item{bsa}{preoperative body surface area.}
}
\description{
Pedict risk scores (decsion values) and  trajectory clusters for a patient (goups of patients)
with  variables:
}
